v.0.9.5 - refactoring
Obsolete methods moved to separate files
Enabled user to specify is session context read-only.
